+++ lib/glimmer-dsl-opal/samples/hello/hello_list_multi_selection.rb
@@ -59,7 +59,7 @@ class HelloListMultiSelection
       text 'Hello, List Multi Selection!'
       
       list(:multi) {
-        selection bind(person, :provinces)
+        selection <=> [person, :provinces]
       }
       
       button {