+++ lib/glimmer/data_binding/table_items_binding.rb
@@ -11,10 +11,11 @@ module Glimmer
       include DataBinding::Observable
       include DataBinding::Observer
 
-      def initialize(parent, model_binding, column_properties)
+      def initialize(parent, model_binding, column_properties = nil)
         @last_populated_model_collection = nil
         @table = parent
         @model_binding = model_binding
+#         @column_properties = model_binding.binding_options[:column_attributes] || model_binding.binding_options[:column_properties] || column_properties # TODO
         @column_properties = column_properties
         @table.data = @model_binding
         ##@table.on_widget_disposed do |dispose_event| # doesn't seem needed within Opal